backend/requirements.txt
anyio==3.6.2
async-timeout==4.0.2
bidict==0.22.1
black==23.1.0
click==8.1.3
fastapi==0.92.0
h11==0.14.0
httptools==0.5.0
idna==3.4
mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.0
platformdirs==3.0.0
pydantic==1.10.5
python-dotenv==0.21.1
python-engineio==4.3.4
python-socketio==5.7.2
PyYAML==6.0
redis==4.5.1
ruff==0.0.247
sniffio==1.3.0
starlette==0.25.0
typing_extensions==4.5.0
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
bcrypt==4.1.1
confluent_kafka==2.3.0
mysql-connector-python==8.2.0